+++ delphyne-demos/demos/scriptlets.py
@@ -199,6 +199,6 @@ def main():
             print("Running simulation for {0} seconds.".format(args.duration))
             stats.start()
             simulation_tree.tick_tock(
-                period=tree_time_step, number_of_iterations=args.duration / tree_time_step
+                period=tree_time_step, number_of_iterations=int(args.duration / tree_time_step)
             )
         launcher.terminate()
